requirements.txt
numpy==1.26.2
pandas>=1.5.0
matplotlib==3.8.2
openpyxl==3.1.2
requests==2.31.0
python-dotenv>=1.0.0
pyyaml==6.0.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
anytree==2.9.1
tqdm==4.66.1
loguru>=0.6.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
selenium==4.17.2
chromedriver-binary
webdriver-manager>=4.0.1
bs4
beautifulsoup4
requests
openai>=1.5.0
tiktoken
google-cloud-storage>=2.6.0
google-cloud-bigquery>=3.3.5